Create CSS file and an HTML file
A CSS file and an HTML file were created and linked together. The
HTML file includes the necessary componants in the boilerplate.
It includes as meta tag with a description and content attribute
to help with SEO. Next step will be creating the structure of
the site in the index.html file I created.