Create basic structure in index.html
Five sections were added to the body of index.html including a
header, 3 divisions for 3 content sections, and a footer. The
header includes a logo and navigation with links. The 3 divisions
include a section for hero text and an image, a section for
displaying the books I am presently reading, and a section for
displaying a quote from one of these books. The footer has a
copyright for practice purposes. Next will involve adding images
and text content in place of what I've used as filler for now.