Adds significant structure and style to project
The main work was in style.css to help style the first of
four sections within the html structure. This section, the
header, now has been styled with flexbox with a navigation bar, logo,
hero text and a button with an image beside them. Background-color
and text color as well as size have been implemented for this section.