Add styles to section with class .action
The .action section needed flexbox. It now has the same padding
as the other sections. It serves as a call to action where
visitors can subscribe to the newsletter for the website. The
max-width property came in handy in positioning the button
and the text portion of this section. It allowed the use of
justify-content property on the .action.text div to create
space between the text section and the button.